[AdminBundle] Moved AssetController under Pimcore\Bundle\AdminBundle\Controller\Admin\Asset - #6415
